+++ custom_components/yamaha_ynca/select.py
@@ -16,7 +16,8 @@ from .const import (
     TWOCHDECODER_STRINGS,
     ZONE_ATTRIBUTE_NAMES,
 )
-from .helpers import YamahaYncaSettingEntity, subunit_supports_entitydescription_key
+from .entity import YamahaYncaSettingEntity
+from .helpers import subunit_supports_entitydescription_key
 
 if TYPE_CHECKING:  # pragma: no cover
     from collections.abc import Callable
